Working properly, calculates whether the trip can be completed

+++ Point.cpp
@@ -4,14 +4,14 @@
 
 #include "Point.h"
 
-Point::Point(int x, int y) : x{x}, y{y} {
+Point::Point(double x, double y) : x{x}, y{y} {
 }
 
 void Point::setX(int value) {
     this->x = value;
 }
 
-int Point::getX() const {
+double Point::getX() const {
     return this->x;
 }
 
@@ -19,6 +19,6 @@ void Point::setY(int value) {
     this->y = value;
 }
 
-int Point::getY() {
+double Point::getY() {
     return this->y;
 }
